Move `log_warning_stats` out of `mcmc` module
This is in preparation of decoupling population-sampling related code
from other functions in `mcmc.py`.
